split description-item into subcomoponents
