+++ src/components/content/DescriptionItem/DescriptionItem.stories.mdx
@@ -2,12 +2,31 @@ import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs'
 import { Theme } from '../../../../.storybook/components'
 import { Section, SectionContainer, SectionContent } from '../../section'
 import { DescriptionItem } from './DescriptionItem'
+import { DescriptionItemContainer } from './DescriptionItemContainer'
+import { DescriptionItemTitle } from './DescriptionItemTitle'
+import { DescriptionItemContent } from './DescriptionItemContent'
 
-<Meta title="Components/Content/DescriptionItem" component={DescriptionItem} />
+<Meta
+  title="Components/Content/DescriptionItem"
+  component={DescriptionItem}
+  subcomponents={{
+    DescriptionItemContainer,
+    DescriptionItemTitle,
+    DescriptionItemContent,
+  }}
+/>
 
 # DescriptionItem
 
 The DescriptionItem displays read only information with HTML description list element (`dl`) tag.
+
+### Subcomponents
+
+It is composed of the following components, which can also be used separately for custom purposes:
+
+- DescriptionItemContainer
+- DescriptionItemTitle
+- DescriptionItemContent
 
 export const Template = ({ title, content, hideIfEmpty }) => (
   <DescriptionItem title={title} content={content} hideIfEmpty={hideIfEmpty} />
